LayoutTests/imported/w3c/web-platform-tests/trusted-types/inheriting-csp-for-local-schemes-expected.txt

CONSOLE MESSAGE: This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: Refused to create a TrustedTypePolicy named 'forbiddenPolicy' because it violates the following Content Security Policy directive: "trusted-types allowedPolicy passThrough"
CONSOLE MESSAGE: This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: Refused to create a TrustedTypePolicy named 'forbiddenPolicy' because it violates the following Content Security Policy directive: "trusted-types allowedPolicy passThrough"
CONSOLE MESSAGE: This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: Refused to create a TrustedTypePolicy named 'forbiddenPolicy' because it violates the following Content Security Policy directive: "trusted-types allowedPolicy passThrough"

PASS require-trusted-types-for directive should be inherited in local srcdoc frames
PASS trusted-types directive should be inherited in local srcdoc frames
PASS require-trusted-types-for directive should be inherited in local data frames
PASS trusted-types directive should be inherited in local data frames
PASS require-trusted-types-for directive should be inherited in local blob frames
FAIL trusted-types directive should be inherited in local blob frames assert_not_equals: got disallowed value null
PASS require-trusted-types-for directive should be inherited in local about:blank frames
PASS trusted-types directive should be inherited in local about:blank frames
PASS require-trusted-types-for directive should not be inherited in local blank.html frames
PASS trusted-types directive should not be inherited in local blank.html frames
